Rename component_registration_utils
