+++ lib/src/react_client/component_registration.dart
